rename: change directory structure
패키지 구조 변경
