+++ src/main/java/com/timcooki/jnuwiki/domain/docs/entity/DocsLocation.java
@@ -3,5 +3,4 @@ package com.timcooki.jnuwiki.domain.docs.entity;
 public class DocsLocation {
     private Double lat;
     private Double lng;
-
 }
