+++ src/main/java/com/timcooki/jnuwiki/util/ApiResult.java
@@ -1,4 +1,4 @@
-package com.timcooki.jnuwiki._core.util;
+package com.timcooki.jnuwiki.util;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
